+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Remote A2A response does not yield duplicate content
 - Added `cds.folders.srvs` to support adjacent agent markdowns ootb.
+- Resets quota counter correctly between tasks for task related quota
 
 ## Version 0.9.3 - 2026-09-04